moved each feature classes to each modules
