+++ common/src/main/java/com/freshdigitable/yttt/feature/subscription/FetchSubscriptionListSourceUseCase.kt
@@ -1,0 +1,8 @@
+package com.freshdigitable.yttt.feature.subscription
+
+import com.freshdigitable.yttt.data.model.LiveSubscription
+import kotlinx.coroutines.flow.Flow
+
+interface FetchSubscriptionListSourceUseCase {
+    operator fun invoke(): Flow<List<LiveSubscription>>
+}
