+++ common/src/main/java/com/freshdigitable/yttt/feature/video/FindLiveVideoUseCase.kt
@@ -1,0 +1,7 @@
+package com.freshdigitable.yttt.feature.video
+
+import com.freshdigitable.yttt.data.model.LiveVideo
+
+interface FindLiveVideoUseCase {
+    suspend operator fun invoke(id: LiveVideo.Id): LiveVideo?
+}
