+++ feature-timetable-twitch/src/main/java/com/freshdigitable/yttt/feature/subscription/FetchSubscriptionListSourceFromTwitchUseCase.kt
@@ -1,27 +1,13 @@
-package com.freshdigitable.yttt
+package com.freshdigitable.yttt.feature.subscription
 
 import com.freshdigitable.yttt.data.TwitchLiveRepository
-import com.freshdigitable.yttt.data.YouTubeRepository
 import com.freshdigitable.yttt.data.model.LiveSubscription
 import com.freshdigitable.yttt.data.model.toLiveSubscription
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 
-interface FetchSubscriptionListSourceUseCase {
-    operator fun invoke(): Flow<List<LiveSubscription>>
-}
-
-class FetchSubscriptionListSourceFromYouTubeUseCase @Inject constructor(
-    private val repository: YouTubeRepository,
-) : FetchSubscriptionListSourceUseCase {
-    override fun invoke(): Flow<List<LiveSubscription>> = flow {
-        val subs = repository.fetchAllSubscribe().map { it.toLiveSubscription() }
-        emit(subs)
-    }
-}
-
-class FetchSubscriptionListSourceFromTwitchUseCase @Inject constructor(
+internal class FetchSubscriptionListSourceFromTwitchUseCase @Inject constructor(
     private val repository: TwitchLiveRepository,
 ) : FetchSubscriptionListSourceUseCase {
     override fun invoke(): Flow<List<LiveSubscription>> = flow {
