+++ feature-timetable-twitch/src/main/java/com/freshdigitable/yttt/feature/video/di/TwitchVideoModule.kt
@@ -1,11 +1,10 @@
-package com.freshdigitable.yttt.di
+package com.freshdigitable.yttt.feature.video.di
 
-import com.freshdigitable.yttt.FindLiveVideoFromTwitchUseCase
-import com.freshdigitable.yttt.FindLiveVideoFromYouTubeUseCase
-import com.freshdigitable.yttt.FindLiveVideoUseCase
 import com.freshdigitable.yttt.data.model.TwitchChannelSchedule
 import com.freshdigitable.yttt.data.model.TwitchStream
-import com.freshdigitable.yttt.data.model.YouTubeVideo
+import com.freshdigitable.yttt.di.IdBaseClassKey
+import com.freshdigitable.yttt.feature.video.FindLiveVideoFromTwitchUseCase
+import com.freshdigitable.yttt.feature.video.FindLiveVideoUseCase
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -15,7 +14,7 @@ import dagger.multibindings.IntoMap
 @Suppress("unused")
 @Module
 @InstallIn(ViewModelComponent::class)
-interface FindLiveVideoModule {
+internal interface TwitchVideoModule {
     @Binds
     @IntoMap
     @IdBaseClassKey(TwitchStream.Id::class)
@@ -25,9 +24,4 @@ interface FindLiveVideoModule {
     @IntoMap
     @IdBaseClassKey(TwitchChannelSchedule.Stream.Id::class)
     fun bindFindLiveVideoFromTwitchUseCaseForStreamScheduleId(useCase: FindLiveVideoFromTwitchUseCase): FindLiveVideoUseCase
-
-    @Binds
-    @IntoMap
-    @IdBaseClassKey(YouTubeVideo.Id::class)
-    fun bindFindLiveVideoFromYouTubeUseCase(useCase: FindLiveVideoFromYouTubeUseCase): FindLiveVideoUseCase
 }
