+++ feature-timetable-youtube/src/main/java/com/freshdigitable/yttt/feature/video/FindLiveVideoFromYouTubeUseCase.kt
@@ -1,0 +1,18 @@
+package com.freshdigitable.yttt.feature.video
+
+import com.freshdigitable.yttt.data.YouTubeFacade
+import com.freshdigitable.yttt.data.model.LiveVideo
+import com.freshdigitable.yttt.data.model.YouTubeVideo
+import com.freshdigitable.yttt.data.model.mapTo
+import com.freshdigitable.yttt.data.model.toLiveVideo
+import javax.inject.Inject
+
+internal class FindLiveVideoFromYouTubeUseCase @Inject constructor(
+    private val facade: YouTubeFacade,
+) : FindLiveVideoUseCase {
+    override suspend fun invoke(id: LiveVideo.Id): LiveVideo? {
+        check(id.type == YouTubeVideo.Id::class)
+        val v = facade.fetchVideoList(listOf(id.mapTo())).firstOrNull()
+        return v?.toLiveVideo()
+    }
+}
