+++ lib/src/main/java/com/freshdigitable/yttt/feature/oauth/TwitchOauthViewModel.kt
@@ -1,4 +1,4 @@
-package com.freshdigitable.yttt
+package com.freshdigitable.yttt.feature.oauth
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
